Initial push/second push
This is my second time pushing this. It bugged the first time and I had to delete and reupload after making some changes and updates. 

This last push, I fixed a bug with the restart button and added a message that displays if you try starting the game before you press a button.